split types into own package
