+++ haskell-ide-engine.cabal
@@ -115,6 +115,7 @@ test-suite haskell-ide-test
                      , fast-logger
                      , haskell-ide-engine
                      , hie-apply-refact
+                     , hie-base
                      , hie-eg-plugin-async
                      , hie-ghc-mod
                      , hie-hare
